+++ index.css
@@ -56,7 +56,7 @@ body {
 
 .books-section {
   margin-top: 3rem;
-  display: block;
+  display: none;
 }
 
 .books-list-title {
@@ -104,7 +104,7 @@ body {
 
 .section-contact {
   margin-top: 3rem;
-  display: block;
+  display: none;
 }
 
 .contact-content {
@@ -136,7 +136,7 @@ body {
 
 .form-section {
   margin-top: 3rem;
-  display: block;
+  display: none;
 }
 
 .form-content {
