+++ modules/addBook.js
@@ -5,8 +5,6 @@ const inputTitle = document.getElementById('title');
 const inputAuthor = document.getElementById('author');
 const errorMesg = document.querySelector('.error-mesg');
 
-
-  
   export default function addBook() {
         if (inputTitle.value !== '' && inputAuthor.value !== '') {
         const book = new Books(inputTitle.value, inputAuthor.value);
